requirements.txt
# Using python 3.10.12
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
g2pkk==0.1.2
grpcio==1.66.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.24.7
idna==3.9
inflect==7.4.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==24.1
pandas==1.5.3
pillow==10.4.0
platformdirs==4.3.3
pooch==1.8.2
preshed==3.0.9
protobuf==5.28.1
psutil==6.0.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pypinyin==0.53.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.4
semantic-version==2.10.0
setuptools-rust==1.10.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.5
trainer==0.0.36
transformers==4.44.2
triton==2.3.1
TTS==0.22.0
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
